+++ mukurobot/__main__.py
@@ -1,3 +1,10 @@
+'''
+Main
+
+(c) 2023 Sakuragasaki46
+See LICENSE for license information
+'''
+
 
 from . import main